refactor(endo): Consolidate archive and import hooks
