+++ packages/endo/src/import.js
@@ -3,71 +3,8 @@
 import { compartmentMapForNodeModules } from "./compartmap.js";
 import { search } from "./search.js";
 import { assemble } from "./assemble.js";
-import { parseExtension } from "./extension.js";
+import { makeImportHookMaker } from "./import-hook.js";
 import * as json from "./json.js";
-
-const decoder = new TextDecoder();
-
-// q, as in quote, for quoting strings in error messages.
-const q = JSON.stringify;
-
-const resolveLocation = (rel, abs) => new URL(rel, abs).toString();
-
-const makeImportHookMaker = (read, baseLocation) => {
-  // per-assembly:
-  const makeImportHook = (packageLocation, parse) => {
-    // per-compartment:
-    packageLocation = resolveLocation(packageLocation, baseLocation);
-    const importHook = async moduleSpecifier => {
-      // per-module:
-
-      // Collate candidate locations for the moduleSpecifier per Node.js
-      // conventions.
-      const candidates = [];
-      if (moduleSpecifier === ".") {
-        candidates.push("index.js");
-      } else {
-        candidates.push(moduleSpecifier);
-        if (parseExtension(moduleSpecifier) === "") {
-          candidates.push(
-            `${moduleSpecifier}.js`,
-            `${moduleSpecifier}/index.js`
-          );
-        }
-      }
-
-      for (const candidate of candidates) {
-        const moduleLocation = resolveLocation(candidate, packageLocation);
-        // eslint-disable-next-line no-await-in-loop
-        const moduleBytes = await read(moduleLocation).catch(
-          _error => undefined
-        );
-        if (moduleBytes !== undefined) {
-          const moduleSource = decoder.decode(moduleBytes);
-
-          const { record } = parse(
-            moduleSource,
-            moduleSpecifier,
-            moduleLocation,
-            packageLocation
-          );
-          return record;
-        }
-      }
-
-      // TODO offer breadcrumbs in the error message, or how to construct breadcrumbs with another tool.
-      throw new Error(
-        `Cannot find file for internal module ${q(
-          moduleSpecifier
-        )} (with candidates ${candidates
-          .map(q)
-          .join(", ")}) in package ${packageLocation}`
-      );
-    };
-    return importHook;
-  };
-  return makeImportHook;
-};
 
 export const loadLocation = async (read, moduleLocation) => {
   const {
